# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Notification:
    properties:
      action_url:
        type: string
      archived_at:
        type: string
      comment:
        $ref: '#/definitions/models.TaskComment'
      comment_id:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      data:
        $ref: '#/definitions/models.NotificationData'
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      message:
        type: string
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      read_at:
        type: string
      status:
        $ref: '#/definitions/models.NotificationStatus'
      task:
        $ref: '#/definitions/models.Task'
      task_id:
        description: Related entity references
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      type:
        $ref: '#/definitions/models.NotificationType'
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relationships
      user_id:
        type: string
    type: object
...
  models.Task:
    properties:
      activities:
        items:
          $ref: '#/definitions/models.TaskActivity'
        type: array
      actual_hours:
        type: number
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      attachments:
        items:
          $ref: '#/definitions/models.TaskAttachment'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      completed_at:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      creator:
        $ref: '#/definitions/models.User'
      creator_id:
        description: Relationships
        type: string
      description:
        type: string
      due_date:
        type: string
      estimated_hours:
        type: number
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.Task'
      parent_id:
        type: string
      priority:
        $ref: '#/definitions/models.TaskPriority'
      project:
        $ref: '#/definitions/models.Project'
      project_id:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      subtasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      title:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  models.TaskActivity:
    properties:
      action:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        type: string
      field:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      new_value:
        type: string
      old_value:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskAttachment:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      mime_type:
        type: string
      original_name:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskComment:
    properties:
      content:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      parent:
        $ref: '#/definitions/models.TaskComment'
      parent_id:
        type: string
      replies:
        items:
          $ref: '#/definitions/models.TaskComment'
        type: array
      task:
        allOf:
        - $ref: '#/definitions/models.Task'
        description: Relationships
      task_id:
        type: string
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TaskPriority:
    enum:
    - low
    - medium
    - high
    - urgent
    type: string
    x-enum-varnames:
    - TaskPriorityLow
    - TaskPriorityMedium
    - TaskPriorityHigh
    - TaskPriorityUrgent
  models.TaskStatus:
    enum:
    - todo
    - in_progress
    - in_review
    - completed
    - canceled
    type: string
    x-enum-varnames:
    - TaskStatusTodo
    - TaskStatusInProgress
    - TaskStatusInReview
    - TaskStatusCompleted
    - TaskStatusCanceled
  models.Tenant:
    properties:
      allow_registration:
        description: Settings fields (flattened)
        type: boolean
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      default_user_role:
        type: string
      domain:
        type: string
      email_notifications:
        description: Notification settings (flattened)
        type: boolean
      favicon_url:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      logo_url:
        type: string
      max_storage:
        description: 1GB in bytes
        type: integer
      max_tasks:
        type: integer
      max_users:
        type: integer
      name:
        type: string
      plan:
        $ref: '#/definitions/models.TenantPlan'
      primary_color:
        description: Branding settings (flattened)
        type: string
      require_email_verification:
        type: boolean
      secondary_color:
        type: string
      slug:
        type: string
      status:
        $ref: '#/definitions/models.TenantStatus'
      task_assignments:
        type: boolean
      task_auto_assignment:
        type: boolean
      task_completions:
        type: boolean
      task_due_dates:
        type: boolean
      tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      users:
        description: Relationships
        items:
          $ref: '#/definitions/models.User'
        type: array
      weekly_digest:
        type: boolean
    type: object
...
  models.User:
    properties:
      assigned_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      avatar:
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      created_tasks:
        items:
          $ref: '#/definitions/models.Task'
        type: array
      default_task_priority:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      email_verified_at:
        type: string
      enable_email_notifications:
        type: boolean
      enable_push_notifications:
        type: boolean
      first_name:
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      language:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      show_completed_tasks:
        type: boolean
      status:
        $ref: '#/definitions/models.UserStatus'
      task_reminders:
        type: boolean
      task_view_mode:
        description: list, board, calendar
        type: string
      tasks_per_page:
        type: integer
      tenant:
        allOf:
        - $ref: '#/definitions/models.Tenant'
        description: Relationships
      tenant_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
      theme:
        description: User preferences (flattened)
        type: string
      timezone:
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      weekly_digest:
        type: boolean
    type: object
...
  models.UserRole:
    enum:
    - admin
    - manager
    - user
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleManager
    - UserRoleUser
  models.UserStatus:
    enum:
    - active
    - inactive
    - suspended
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusSuspended
  requests.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  requests.UpdateUserPreferencesRequest:
    properties:
      default_task_priority:
        enum:
        - low
        - medium
        - high
        - urgent
        type: string
      email_notifications:
        type: boolean
      push_notifications:
        type: boolean
      show_completed_tasks:
        type: boolean
      task_reminders:
        type: boolean
      task_view_mode:
        enum:
        - list
        - board
        - calendar
        type: string
      tasks_per_page:
        maximum: 100
        minimum: 5
        type: integer
      theme:
        enum:
        - light
        - dark
        - auto
        type: string
      weekly_digest:
        type: boolean
    type: object
  response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@taskflow.com
    name: API Support
    url: http://www.taskflow.com/support
  description: A multi-tenant task management platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TaskFlow API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get a paginated list of users in the current tenant
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft delete a user (requires admin role)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a specific user by ID within the current tenant
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user's information (requires admin or manager role)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/stats:
    get:
      description: Get statistics for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user statistics
      tags:
      - users
  /users/change-password:
    post:
      consumes:
      - application/json
      description: Change password for the current user
      parameters:
      - description: Password change data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
  /users/preferences:
    put:
      consumes:
      - application/json
      description: Update preferences for the current user
      parameters:
      - description: User preferences
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserPreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user preferences
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"